Rest in progress

+++ server/src/main/java/cz/cvut/fel/rsp/server/service/RegUserService.java
@@ -5,6 +5,7 @@
  */
 package cz.cvut.fel.rsp.server.service;
 
+import cz.cvut.fel.rsp.server.Model.Enums.UserRoleEnum;
 import cz.cvut.fel.rsp.server.Model.Hotel;
 import cz.cvut.fel.rsp.server.Model.Reservation;
 import cz.cvut.fel.rsp.server.Model.Room;
@@ -15,6 +16,8 @@ import cz.cvut.fel.rsp.server.dao.RoomDao;
 import cz.cvut.fel.rsp.server.dao.UnregisteredUserDao;
 import cz.cvut.fel.rsp.server.dao.UserDao;
 import java.util.List;
+import javax.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 /**
@@ -23,44 +26,75 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class RegUserService extends DaoConnection {
-    
+
+    @Autowired
     public RegUserService(HotelDao hotelDao, UserDao userDao, ReservationDao resDao, UnregisteredUserDao unregUserDao, RoomDao roomDao) {
         super(hotelDao, userDao, resDao, unregUserDao, roomDao);
     }
-    
+
+    @Transactional
     public List<User> findAll() {
         return userDao.findAll();
     }
-    
-    public void createUser(User u, Hotel h) {
+
+    @Transactional
+    public boolean createUser(User u, int hotelId) {
+        Hotel h = hotelDao.find(hotelId);
+        if(h == null) {
+            return false;
+        }
         u.setHotel(h);
         h.getRegisteredUsers().add(u);
         userDao.persist(u);
         hotelDao.update(h);
+        return true;
+    }
+
+    @Transactional
+    public boolean updateUser(User u) {
+        User old = userDao.find(u.getId());
+        if (old != null) {
+            u.setReservations(old.getReservations());
+            u.setRoles(u.getRoles());
+            userDao.update(u);
+            return true;
+        }
+        return false;
     }
     
-    public void updateUser(User u) {
-        userDao.update(u);
+    @Transactional
+    public boolean updateUserRoles(int id, List<UserRoleEnum> roles) {
+        User u = userDao.find(id);
+        if(u != null) {
+            u.setRoles(roles);
+            return true;
+        }
+        return false;
     }
-    
-    public boolean deleteUser(User u) {
-        if(!u.isDeleteable()) {
+
+    @Transactional
+    public boolean deleteUser(int id) {
+        User u = userDao.find(id);
+        if (u == null) {
+            return false;
+        }
+        if (!u.isDeleteable()) {
             return false;
         }
         List<Reservation> reservations = u.getReservations();
         Hotel h = u.getHotel();
-        for(Reservation r: reservations) {
+        for (Reservation r : reservations) {
             List<Room> rooms = r.getRooms();
-            for(Room room: rooms) {
+            for (Room room : rooms) {
                 room.getReservations().remove(r);
                 roomDao.update(room);
             }
             h.getReservations().remove(r);
-            resDao.remove(r);            
+            resDao.remove(r);
         }
         hotelDao.update(h);
         userDao.remove(u);
         return true;
     }
-    
+
 }
